+++ src/utils/password_generator.py
@@ -61,7 +61,7 @@ class DicewarePassphrase:
         Returns:
             Capitalized word as bytes.
         """
-        return bytearray(self._words[key].encode(UTF8))
+        return bytearray(self._words[key].capitalize().encode(UTF8))
 
     def clear(self):
         """Clear loaded wordlist."""
@@ -320,18 +320,15 @@ def ask_password(prompt: str = "Password:") -> bytearray | None:
             pw = bytearray()
             for i in range(pw_len):
                 word = diceware.get_word(diceware.roll_5dice())
-                word = apply_random_capitalization(word)
+                word = apply_random_capitalization(word.decode(UTF8))
                 if use_nums:
                     word = apply_random_number(word)
 
-                pw.extend(word)
-
+                pw.extend(bytearray(word, UTF8))
                 if i < pw_len-1:
                     pw.extend(PASS_DEFAULTS["phrase_sep"][randbelow_reject(
                         len(PASS_DEFAULTS["phrase_sep"]))].encode(UTF8))
-                    
                 del word
-                
                 
             print(f"\n Generated: ", end="", flush= True)
             print_bytearray(pw)
@@ -373,72 +370,26 @@ def check_max_consecutive_bytes(pw: bytearray, threshold: int) -> bool:
 
     return True
 
-def apply_random_capitalization(word: bytearray) -> bytearray:
+def apply_random_capitalization(word: str) -> str:
     r = randbelow_reject(3)
-    for i, b in enumerate(word):
-        if 0x41 <= b <= 0x5A or 0x61 <= b <= 0x7A:
-            # Lowercase, bitwise OR with capital bit. Adds capital bit (0010 0000)
-            if r == 0:
-                word[i] = b | 0x20
-            # Capitalize, bitwise AND with compliment of capital bit. Removes capital bit
-            elif r == 1:
-                if i == 0:
-                    word[i] = b & ~0x20
-                else:
-                    word[i] = b | 0x20
-            # Uppercase
-            else:
-                word[i] = b & ~0x20
-    return word
-
-def apply_random_number(word: bytearray) -> bytearray:
+    if r == 0:
+        return word.lower()
+    elif r == 1:
+        return word.capitalize()
+    else:
+        return word.upper()
+    
+def apply_random_number(word: str) -> str:
     r = randbelow_reject(4)
-
     if r == 0:
         return word
-
-    n = randbelow_reject(10)
-    num = int_to_ascii_ba(n)
-
-    if r == 1:
-        # word + number
-        out = bytearray(len(word) + len(num))
-        out[:len(word)] = word
-        out[len(word):] = num
-
+    elif r == 1:
+        return word \
+                + str(randbelow_reject(50) + randbelow_reject(50))
     elif r == 2:
-        # number + word
-        out = bytearray(len(word) + len(num))
-        out[:len(num)] = num
-        out[len(num):] = word
-
-    else:
-        # number + word + number
-        n2 = randbelow_reject(10)
-        num2 = int_to_ascii_ba(n2)
-
-        out = bytearray(len(num) + len(word) + len(num2))
-        i = 0
-        out[i:i+len(num)] = num
-        i += len(num)
-        out[i:i+len(word)] = word
-        i += len(word)
-        out[i:i+len(num2)] = num2
-
-    return out
-    
-def int_to_ascii_ba(n: int) -> bytearray:
-    if n == 0:
-        return bytearray(b"0")
-
-    out = bytearray()
-    while n > 0:
-        out.append(0x30 + (n % 10))  # '0' + digit
-        n //= 10
-
-    out.reverse()
-    return out
-
-def wipe_byte_arr(buf):
-    for i in range(len(buf)):
-        buf[i] = 0
+        return  str(randbelow_reject(50) + randbelow_reject(50))\
+                + word
+    elif r == 3:
+        return  str(randbelow_reject(50) + randbelow_reject(50)) \
+                + word \
+                + str(randbelow_reject(50) + randbelow_reject(50))